Refactor variation attributes handling: remove old implementation and introduce new structure
